net/bnxt: fix mismatched type comparison in Rx
[ upstream commit 0797fcb00426bf246e83caf074822b4427ec3937 ]

Fix comparison between uint16_t and uint32_t types.

Fixes: 6dc83230b43b ("net/bnxt: support port representor data path")

+++ drivers/net/bnxt/bnxt_rxr.c
@@ -966,7 +966,7 @@ uint16_t bnxt_recv_pkts(void *rx_queue, struct rte_mbuf **rx_pkts,
 
 	/* Attempt to alloc Rx buf in case of a previous allocation failure. */
 	if (alloc_failed) {
-		uint16_t cnt;
+		int cnt;
 
 		for (cnt = 0; cnt < nb_rx_pkts + nb_rep_rx_pkts; cnt++) {
 			struct rte_mbuf **rx_buf;
